+++ ECom.DataModel/Category.cs
@@ -2,6 +2,7 @@
 {
     using System;
     using System.Collections.Generic;
+    using System.ComponentModel.DataAnnotations;
     using System.Linq;
     using System.Text;
     using System.Threading.Tasks;
@@ -14,11 +15,14 @@
         /// <summary>
         /// Gets or Sets the CategoryID
         /// </summary>
+        [Key]
         public int CategoryID { get; set; }
         
         /// <summary>
         /// Gets or Sets the CategoryName
         /// </summary>
         public string CategoryName { get; set; }
+
+        public IEnumerable<Product> Products { get; set; }
     }
 }
