+++ ECom.DataModel/ECom.DataModel.csproj
@@ -7,8 +7,8 @@
     <ProjectGuid>{6E1AF8D0-A04C-474A-96D3-7D7601C35CEB}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>Repositories.DataModel</RootNamespace>
-    <AssemblyName>Repositories.DataModel</AssemblyName>
+    <RootNamespace>ECom.DataModel</RootNamespace>
+    <AssemblyName>ECom.DataModel</AssemblyName>
     <TargetFrameworkVersion>v4.5.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
   </PropertyGroup>
@@ -30,7 +30,16 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
+    <Reference Include="EntityFramework, Version=6.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089, processorArchitecture=MSIL">
+      <HintPath>..\packages\EntityFramework.6.1.3\lib\net45\EntityFramework.dll</HintPath>
+      <Private>True</Private>
+    </Reference>
+    <Reference Include="EntityFramework.SqlServer, Version=6.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089, processorArchitecture=MSIL">
+      <HintPath>..\packages\EntityFramework.6.1.3\lib\net45\EntityFramework.SqlServer.dll</HintPath>
+      <Private>True</Private>
+    </Reference>
     <Reference Include="System" />
+    <Reference Include="System.ComponentModel.DataAnnotations" />
     <Reference Include="System.Core" />
     <Reference Include="System.Xml.Linq" />
     <Reference Include="System.Data.DataSetExtensions" />
@@ -43,11 +52,25 @@
     <Compile Include="Address.cs" />
     <Compile Include="Category.cs" />
     <Compile Include="Customer.cs" />
+    <Compile Include="EcomContext.cs" />
+    <Compile Include="Migrations\201606221303263_InitialCreate.cs" />
+    <Compile Include="Migrations\201606221303263_InitialCreate.Designer.cs">
+      <DependentUpon>201606221303263_InitialCreate.cs</DependentUpon>
+    </Compile>
+    <Compile Include="Migrations\Configuration.cs" />
     <Compile Include="PriceTag.cs" />
     <Compile Include="Product.cs" />
-    <Compile Include="Order.cs" />
     <Compile Include="ProductImage.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="App.config" />
+    <None Include="packages.config" />
+  </ItemGroup>
+  <ItemGroup>
+    <EmbeddedResource Include="Migrations\201606221303263_InitialCreate.resx">
+      <DependentUpon>201606221303263_InitialCreate.cs</DependentUpon>
+    </EmbeddedResource>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
