+++ ECom.DataModel/PriceTag.cs
@@ -1,19 +1,20 @@
 ﻿namespace Ecom.DataModel
 {
     using System;
-    /// <summary>
-    /// Class handles the Default price of Product
-    /// </summary>
+    using System.ComponentModel.DataAnnotations;    /// <summary>
+                                                    /// Class handles the Default price of Product
+                                                    /// </summary>
     public class PriceTag
     {
         /// <summary>
         /// Gets or Sets the UniqueId for unit price of a product.
         /// </summary>
+        [Key]
         public int PriceTagId { get; set; }
         /// <summary>
         /// Gets or Sets the Product Id for mapping to unit price.
         /// </summary>
-        public int ProductId { get; set; }
+        public Product ProductId { get; set; }
         /// <summary>
         /// Gets or Sets The unit price of the product
         /// </summary>
@@ -26,5 +27,7 @@
         /// Gets or Sets whether the price is current item price
         /// </summary>
         public bool IsCurrentRate { get; set; }
+
+        public IEquatable<Product> Products { get; set; }
     }
 }
