+++ ECom.DataModel/ProductImage.cs
@@ -1,5 +1,6 @@
 ﻿namespace Ecom.DataModel
 {
+    using System.ComponentModel.DataAnnotations;
     /// <summary>
     /// Class regarding the Image of Product
     /// </summary>
@@ -8,11 +9,12 @@
         /// <summary>
         /// Gets or Sets the ImageIf
         /// </summary>
+        [Key]
         public int ImageId { get; set; }
         /// <summary>
         /// Gets or Sets the Mapping productId
         /// </summary>
-        public int ProductId { get; set; }
+        public Product ProductId { get; set; }
         /// <summary>
         /// Gets or Sets the Image Name or Unique name
         /// </summary>
@@ -21,6 +23,10 @@
         /// Gets or Sets if the Image is update or not.
         /// </summary>
         public bool IsActiveImage { get; set; }
+        /// <summary>
+        /// Gets or Sets the Image order when it is shown to user
+        /// </summary>
+        public int OrderOfImage { get; set; }
 
     }
 }
