+++ ECom.DataModel/Properties/AssemblyInfo.cs
