+++ Ecom/Properties/AssemblyInfo.cs
