--- Repositories.DataModel/Order.cs
@@ -1,17 +1,0 @@
-﻿namespace Ecom.DataModel
-{
-    using System;
-    using System.Collections.Generic;
-    using System.Linq;
-    using System.Text;
-    using System.Threading.Tasks;
-
-    public class Order
-    {
-        public int OrderId { get; set; }
-
-        public int CustomerId { get; set; }
-
-
-    }
-}
